requirements.txt
# VirtualOutfit AI - Python Backend
# Image Generation Pipeline with FASHN.ai

# Requirements for this backend
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
python-dotenv>=1.0.0
httpx>=0.25.0



